+++ glomar_gridding/ellipse/__init__.py
@@ -1,0 +1,12 @@
+"""Covariance Estimation using Ellipses"""
+
+from glomar_gridding.ellipse.model import EllipseModel
+from glomar_gridding.ellipse.estimate import EllipseBuilder
+from glomar_gridding.ellipse.covariance import EllipseCovarianceBuilder
+
+
+__all__ = [
+    "EllipseBuilder",
+    "EllipseCovarianceBuilder",
+    "EllipseModel",
+]